Update LiteX packages to the latest revisions
This commit updates the LiteX package set to the latest revisions,
as of 2024-01-09. This update features a couple regressions.

With this update, the VexRiscv SMP CPU is still built, but no longer
tested in the LiteX CPU tests. This is because LiteX now builds
missing CPU targets as part of its tests, and the LiteX test closure
deliberately does not (yet) contain all packages required for testing
these CPU variants. This can either be fixed by including all required
packages in the LiteX package's nativeCheckInputs, or by determining
the exact set of CPU variants required by the LiteX tests and building
them ahead of time in the VexRiscv SMP derivation, in addition to all
the other CPU variants included in its Makefile.

Furthermore, some LiteDRAM tests fail after this update. This is
because of an incompatibility with the Migen version shipped on recent
NixOS / Nixpkgs releases, in particular `unstable-2022-09-02` / git
rev `639e66f4f45343`. The Migen version used in the LiteX CI itself
(v0.9.2) in turn fails to run some of its tests (and subsequently
errors in the LiteX tests) on more recent Python versions (works on
3.9, fails on 3.11). The following override can be added to the Python
overlay to swtich to this Migen version, if a sufficiently old Python
version is used:

    migen = super.migen.overridePythonAttrs (oldAttrs: {
      # Unfortunately, some litedram tests fail when used with more recent
      # versions of Migen (e.g., unstable-2022-09-02 as shipped with NixOS
      # 23.11, rev 639e66f4f45343). On sufficiently old Python versions (e.g.,
      # 3.9) we can use the Migen release 0.9.2, as pinned in the LiteX CI:
      src = pkgs.fetchFromGitHub {
        owner = "m-labs";
        repo = "migen";
        rev = "0.9.2";
        sha256 = "sha256-9Bj/Qh0/5BVbYrBsLmiX/YTrTKAHB072I0h+YlwMAc8=";
      };

      # The Migen tests themselves fail on this revision with a sufficiently
      # recent version of Python (working on 3.9, raises ValueError on Python
      # 3.11 with message "Cannot extract clock domain name from code, need to
      # specify."). Uncomment the following to skip the affected test.
      #nativeCheckInputs = (oldAttrs.checkInputs or []) ++ [ super.pytest ];
      #checkPhase = ''pytest -v -k "not test_local_cd"'';
    });

This issue will ultimately have to be fixed upstream, in either LiteX
or Migen. For now, it seems that only the LiteDRAM tests itself are
affected, and at least the set of `litex-boards` tests using LiteDRAM
still pass.

In addition to these changes, upstream LiteX added support for the
Microwatt and NEORV32 CPUs, which are still unsupported within
nix-litex and hence disabled in tests.

Signed-off-by: Leon Schuermann <leon@is.currently.online>